# Speed up replaceRegexpOne/All with vectorscan 