--- a/src/Functions/ReplaceRegexpImpl.h
+++ b/src/Functions/ReplaceRegexpImpl.h
@@ -1,7 +1,6 @@
 #pragma once
 
 #include <base/types.h>
-#include <Common/Volnitsky.h>
 #include <Columns/ColumnString.h>
 #include <IO/WriteHelpers.h>
 
@@ -17,131 +16,130 @@ namespace ErrorCodes
     extern const int BAD_ARGUMENTS;
 }
 
+struct ReplaceRegexpTraits
+{
+    enum class Replace
+    {
+        First,
+        All
+    };
+};
 
 /** Replace all matches of regexp 'needle' to string 'replacement'. 'needle' and 'replacement' are constants.
-  * 'replacement' could contain substitutions, for example: '\2-\3-\1'
+  * 'replacement' can contain substitutions, for example: '\2-\3-\1'
   */
-template <bool replace_one = false>
+template <ReplaceRegexpTraits::Replace replace>
 struct ReplaceRegexpImpl
 {
-    /// Sequence of instructions, describing how to get resulting string.
     struct Instruction
     {
-        /// If not negative - perform substitution of n-th subpattern from the regexp match.
+        /// If not negative, perform substitution of n-th subpattern from the regexp match.
         int substitution_num = -1;
-        /// Otherwise - paste this string verbatim.
-        std::string literal;
+        /// Otherwise, paste this literal string verbatim.
+        String literal;
 
-        Instruction(int substitution_num_) : substitution_num(substitution_num_) {} /// NOLINT
-        Instruction(std::string literal_) : literal(std::move(literal_)) {} /// NOLINT
+        explicit Instruction(int substitution_num_) : substitution_num(substitution_num_) {}
+        explicit Instruction(String literal_) : literal(std::move(literal_)) {}
     };
 
+    /// Decomposes the replacement string into a sequence of substitutions and literals.
+    /// E.g. "abc\1de\2fg\1\2" --> inst("abc"), inst(1), inst("de"), inst(2), inst("fg"), inst(1), inst(2)
     using Instructions = std::vector<Instruction>;
 
-    static const size_t max_captures = 10;
+    static constexpr int max_captures = 10;
 
-
-    static Instructions createInstructions(const std::string & s, int num_captures)
+    static Instructions createInstructions(std::string_view replacement, int num_captures)
     {
         Instructions instructions;
 
-        String now;
-        for (size_t i = 0; i < s.size(); ++i)
+        String literals;
+        for (size_t i = 0; i < replacement.size(); ++i)
         {
-            if (s[i] == '\\' && i + 1 < s.size())
+            if (replacement[i] == '\\' && i + 1 < replacement.size())
             {
-                if (isNumericASCII(s[i + 1])) /// Substitution
+                if (isNumericASCII(replacement[i + 1])) /// Substitution
                 {
-                    if (!now.empty())
+                    if (!literals.empty())
                     {
-                        instructions.emplace_back(now);
-                        now = "";
+                        instructions.emplace_back(literals);
+                        literals = "";
                     }
-                    instructions.emplace_back(s[i + 1] - '0');
+                    instructions.emplace_back(replacement[i + 1] - '0');
                 }
                 else
-                    now += s[i + 1]; /// Escaping
+                    literals += replacement[i + 1]; /// Escaping
                 ++i;
             }
             else
-                now += s[i]; /// Plain character
+                literals += replacement[i]; /// Plain character
         }
 
-        if (!now.empty())
-        {
-            instructions.emplace_back(now);
-            now = "";
-        }
+        if (!literals.empty())
+            instructions.emplace_back(literals);
 
-        for (const auto & it : instructions)
-            if (it.substitution_num >= num_captures)
-                throw Exception(ErrorCodes::BAD_ARGUMENTS,
-                    "Invalid replace instruction in replacement string. Id: {}, but regexp has only {} subpatterns",
-                    it.substitution_num, num_captures - 1);
+        for (const auto & instr : instructions)
+            if (instr.substitution_num >= num_captures)
+                throw Exception(
+                    ErrorCodes::BAD_ARGUMENTS,
+                    "Id {} in replacement string is an invalid substitution, regexp has only {} capturing groups",
+                    instr.substitution_num, num_captures - 1);
 
         return instructions;
     }
 
-
     static void processString(
-        const re2_st::StringPiece & input,
+        const char * haystack_data,
+        size_t haystack_length,
         ColumnString::Chars & res_data,
         ColumnString::Offset & res_offset,
-        re2_st::RE2 & searcher,
+        const re2_st::RE2 & searcher,
         int num_captures,
         const Instructions & instructions)
     {
+        re2_st::StringPiece haystack(haystack_data, haystack_length);
         re2_st::StringPiece matches[max_captures];
 
         size_t copy_pos = 0;
         size_t match_pos = 0;
 
-        while (match_pos < static_cast<size_t>(input.length()))
+        while (match_pos < haystack_length)
         {
             /// If no more replacements possible for current string
             bool can_finish_current_string = false;
 
-            if (searcher.Match(input, match_pos, input.length(), re2_st::RE2::Anchor::UNANCHORED, matches, num_captures))
+            if (searcher.Match(haystack, match_pos, haystack_length, re2_st::RE2::Anchor::UNANCHORED, matches, num_captures))
             {
-                const auto & match = matches[0];
-                size_t bytes_to_copy = (match.data() - input.data()) - copy_pos;
+                const auto & match = matches[0]; /// Complete match (\0)
+                size_t bytes_to_copy = (match.data() - haystack.data()) - copy_pos;
 
-                /// Copy prefix before matched regexp without modification
+                /// Copy prefix before current match without modification
                 res_data.resize(res_data.size() + bytes_to_copy);
-                memcpySmallAllowReadWriteOverflow15(&res_data[res_offset], input.data() + copy_pos, bytes_to_copy);
+                memcpySmallAllowReadWriteOverflow15(&res_data[res_offset], haystack.data() + copy_pos, bytes_to_copy);
                 res_offset += bytes_to_copy;
                 copy_pos += bytes_to_copy + match.length();
                 match_pos = copy_pos;
 
-                /// Do substitution instructions
-                for (const auto & it : instructions)
+                /// Substitute inside current match using instructions
+                for (const auto & instr : instructions)
                 {
-                    if (it.substitution_num >= 0)
-                    {
-                        const auto & substitution = matches[it.substitution_num];
-
-                        res_data.resize(res_data.size() + substitution.length());
-                        memcpy(&res_data[res_offset], substitution.data(), substitution.length());
-                        res_offset += substitution.length();
-                    }
+                    std::string_view replacement;
+                    if (instr.substitution_num >= 0)
+                        replacement = std::string_view(matches[instr.substitution_num].data(), matches[instr.substitution_num].size());
                     else
-                    {
-                        const auto & literal = it.literal;
-
-                        res_data.resize(res_data.size() + literal.size());
-                        memcpy(&res_data[res_offset], literal.data(), literal.size());
-                        res_offset += literal.size();
-                    }
+                        replacement = instr.literal;
+                    res_data.resize(res_data.size() + replacement.size());
+                    memcpy(&res_data[res_offset], replacement.data(), replacement.size());
+                    res_offset += replacement.size();
                 }
 
-                if (replace_one)
+                if constexpr (replace == ReplaceRegexpTraits::Replace::First)
                     can_finish_current_string = true;
 
-                if (match.length() == 0)
+                if (match.empty())
                 {
                     /// Step one character to avoid infinite loop
                     ++match_pos;
-                    if (match_pos >= static_cast<size_t>(input.length()))
+                    if (match_pos >= haystack_length)
                         can_finish_current_string = true;
                 }
             }
@@ -151,10 +149,10 @@ struct ReplaceRegexpImpl
             /// If ready, append suffix after match to end of string.
             if (can_finish_current_string)
             {
-                res_data.resize(res_data.size() + input.length() - copy_pos);
-                memcpySmallAllowReadWriteOverflow15(&res_data[res_offset], input.data() + copy_pos, input.length() - copy_pos);
-                res_offset += input.length() - copy_pos;
-                copy_pos = input.length();
+                res_data.resize(res_data.size() + haystack_length - copy_pos);
+                memcpySmallAllowReadWriteOverflow15(&res_data[res_offset], haystack.data() + copy_pos, haystack_length - copy_pos);
+                res_offset += haystack_length - copy_pos;
+                copy_pos = haystack_length;
                 match_pos = copy_pos;
             }
         }
@@ -164,12 +162,11 @@ struct ReplaceRegexpImpl
         ++res_offset;
     }
 
-
     static void vector(
         const ColumnString::Chars & data,
         const ColumnString::Offsets & offsets,
-        const std::string & needle,
-        const std::string & replacement,
+        const String & needle,
+        const String & replacement,
         ColumnString::Chars & res_data,
         ColumnString::Offsets & res_offsets)
     {
@@ -178,11 +175,19 @@ struct ReplaceRegexpImpl
         size_t size = offsets.size();
         res_offsets.resize(size);
 
-        typename re2_st::RE2::Options regexp_options;
-        /// Never write error messages to stderr. It's ignorant to do it from library code.
+        re2_st::RE2::Options regexp_options;
+        /// Don't write error messages to stderr.
         regexp_options.set_log_errors(false);
+
         re2_st::RE2 searcher(needle, regexp_options);
-        int num_captures = std::min(searcher.NumberOfCapturingGroups() + 1, static_cast<int>(max_captures));
+
+        if (!searcher.ok())
+            throw Exception(
+                ErrorCodes::BAD_ARGUMENTS,
+                "The pattern argument is not a valid re2 pattern: {}",
+                searcher.error());
+
+        int num_captures = std::min(searcher.NumberOfCapturingGroups() + 1, max_captures);
 
         Instructions instructions = createInstructions(replacement, num_captures);
 
@@ -190,9 +195,10 @@ struct ReplaceRegexpImpl
         for (size_t i = 0; i < size; ++i)
         {
             size_t from = i > 0 ? offsets[i - 1] : 0;
-            re2_st::StringPiece input(reinterpret_cast<const char *>(data.data() + from), offsets[i] - from - 1);
+            const char * haystack_data = reinterpret_cast<const char *>(data.data() + from);
+            const size_t haystack_length = static_cast<unsigned>(offsets[i] - from - 1);
 
-            processString(input, res_data, res_offset, searcher, num_captures, instructions);
+            processString(haystack_data, haystack_length, res_data, res_offset, searcher, num_captures, instructions);
             res_offsets[i] = res_offset;
         }
     }
@@ -200,8 +206,8 @@ struct ReplaceRegexpImpl
     static void vectorFixed(
         const ColumnString::Chars & data,
         size_t n,
-        const std::string & needle,
-        const std::string & replacement,
+        const String & needle,
+        const String & replacement,
         ColumnString::Chars & res_data,
         ColumnString::Offsets & res_offsets)
     {
@@ -210,20 +216,29 @@ struct ReplaceRegexpImpl
         res_data.reserve(data.size());
         res_offsets.resize(size);
 
-        typename re2_st::RE2::Options regexp_options;
-        /// Never write error messages to stderr. It's ignorant to do it from library code.
+        re2_st::RE2::Options regexp_options;
+        /// Don't write error messages to stderr.
         regexp_options.set_log_errors(false);
+
         re2_st::RE2 searcher(needle, regexp_options);
-        int num_captures = std::min(searcher.NumberOfCapturingGroups() + 1, static_cast<int>(max_captures));
+
+        if (!searcher.ok())
+            throw Exception(
+                ErrorCodes::BAD_ARGUMENTS,
+                "The pattern argument is not a valid re2 pattern: {}",
+                searcher.error());
+
+        int num_captures = std::min(searcher.NumberOfCapturingGroups() + 1, max_captures);
 
         Instructions instructions = createInstructions(replacement, num_captures);
 
         for (size_t i = 0; i < size; ++i)
         {
             size_t from = i * n;
-            re2_st::StringPiece input(reinterpret_cast<const char *>(data.data() + from), n);
+            const char * haystack_data = reinterpret_cast<const char *>(data.data() + from);
+            const size_t haystack_length = n;
 
-            processString(input, res_data, res_offset, searcher, num_captures, instructions);
+            processString(haystack_data, haystack_length, res_data, res_offset, searcher, num_captures, instructions);
             res_offsets[i] = res_offset;
         }
     }