… docs/en/sql-reference/functions/string-replace-functions.md
… src/Functions/replaceAll.cpp
… src/Functions/replaceOne.cpp
… src/Functions/replaceRegexpAll.cpp
… src/Functions/replaceRegexpOne.cpp