This PR optimizes SQL function `replaceRegexpAll` for the common case that the replacement string contains no substitutions `\1 ... \9` (the trivial substitution `\0` is allowed). In that case, we now evaluate the input using vectorscan instead of re2. Inspiration for this came from https://github.com/StarRocks/starrocks/pull/10028, but here we take it even further and don't limit ourselves to constant patterns.

Note for reviewers: Instead of reviewing everything, it makes more sense to review the (few) individual commits individually.

Benchmark results:

``` sql
-- Generated test data:
insert into test select number, concat('prefix ', toString(toDateTime64('1970-01-01', 3) + number), ' midfix ', toString(toDateTime64('1970-01-01', 3) + number), ' suffix') as d FROM numbers(50000000);

--

┌─id───────┬─haystack─────────────────────────────────────────────────────────────┐
│  0       │ prefix 1970-01-01 00:00:00.000 midfix 1970-01-01 00:00:00.000 suffix │
│  1       │ prefix 1970-01-01 00:00:01.000 midfix 1970-01-01 00:00:01.000 suffix │
│  2       │ prefix 1970-01-01 00:00:02.000 midfix 1970-01-01 00:00:02.000 suffix │

 ... ~50 mil. rows ...

│ 49990001 │ prefix 1971-08-02 14:06:41.000 midfix 1971-08-02 14:06:41.000 suffix │
│ 49990000 │ prefix 1971-08-02 14:06:40.000 midfix 1971-08-02 14:06:40.000 suffix │
└─────────────────────────────────────────────────────────────────────────────────┘
```

Test query:
```sql
select replaceRegexpAll(haystack, pattern, replacement) from test SETTINGS allow_hyperscan=0/1
```

pattern | replacement | allow_hyperscan = 0  | allow_hyperscan = 1 (before tuning) | allow_hyperscan (after tuning) | comment |
| ------ | ------------ | ---------------------- | ------------------- | ---------- | --- |
| 'abc' | 'X' | 1.67 s | 1.68 s | 1.67 s | non-regex, matches nothing |
| 'midfix' | 'X' | 1.55 s | 1.56 s | 1.56 s | non-regex, matches all, simple replacement |
| 'midfix' | '\\0 XYZ \\0 ZXY \\0 YZX \\0' | 2.3 s | 2.3 s | 2.27 s | non-regex, matches all, complex replacement |
| '[xifdim]' | 'X' | 5.3 s | 3.4 s | 4.9 s | regexp (character class), matches all, simple replacement |
| '(pre\|mid\|suf)fix' | 'X' | 1.7 s | 1.7 s | 2.1 s | regexp (disjunction), matches all, simple replacement |
| '197.*fix' | 'X' | 1.2 s | 2.0 s | 2.1 s | regexp (kleene star), matches all, simple replacement |
| '0*' | 'X' | 22 s | "Start of match is not currently supported for patterns which match an empty buffer" | "Start of match is not currently supported for patterns which match an empty buffer" | regexp (kleene star), matches all, simple replacement |
| '[^t]' | 'X' | 23.3 s | 21.2 s | 17.1 s | regexp (negated character class), matches all, simple replacement |
| '197(0\|1)-\d\d-.* \d\d:\d\d:\d\d.000 midfix' | 'X' | 1.6 s | 1.6 s | 1.7 s | complex regexp, matches all, simple replacement |

Before tuning = 2nd commit.

After tuning = 3rd commit (compile vectorscan database and allocate scratch space just once per predicate instead of once per chunk, buffer matches in hash map instead of red/black tree).

### Changelog category (leave one):
- Performance Improvement

### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Speed up SQL function "replaceRegexpAll" for the common case that the replacement string contains no substitutions "\1" - "\9".