@rschu1ze 2022-11-02
After spending some time iterating on this optimization, I am not really happy with the results.

[The StarRock](https://github.com/StarRocks/starrocks/pull/10028) uses vectorscan instead of re2 for constant patterns and they report speedups of 2x-3x. In my measurements, there was no difference for constant patterns at all and why would there be any?

Here, I tried to use vectorscan instead of re2 for patterns without non-trivial capturing group (\1-\9). This is a quite common case but there are some obstacles to performance:
1. Replacing matches requires knowledge of where the match starts. In vectorscan, this is achieved by `HS_FLAG_SOM_LEFTMOST`. Unfortunately, this a) incurs a performance hit according to vectorscan docs b) rejects some valid patterns (e.g. `0*` --> "Start of match is not currently supported for patterns which match an empty buffer" - we'd need some fallback to re2, c) disallows usage of the internal compiled regex cache since all other users of vectorscan in ClickHouse omit `HS_FLAG_SOM_LEFTMOST`.
2. Unlike re2, vectorscan matches are not guaranteed to be reported in order. That doesn't matter if one is only interested in the existence of a match (which is the case for all other vectorscan usages in ClickHouse right now). It means that we cannot assemble the result string on the go as with re2 but need to collect all matches, sort them and assemble the result string ... which deteriorates performance. It also means that `replaceRegexOne()` no longer has a performance advantage over `replaceRegexAll()` (I disabled it for for vectorscan).
3. Finally, re2 does greedy (rightmost) matching whereas vectorscan reports all matches at a given position, starting with the lazy (leftmost) match. There seems no way to change this behavior. So to achieve the same result as with re2, the code needs to discard all but the longest match which again deteriorates performance.

With all the workarounds implemented, benchmark results are mixed. I don't think there is a reliable way to decide by looking at the pattern if evaluation with re2 or vectorscan will be faster.

I am a bit afraid of other subtle differences in matching (even though I am not aware of any). This stopped me from implementing a fallback to re2 when vectorscan rejects a pattern (it would be easy to build), and I also think evaluation with vectorscan for `replaceRegexAll()` should be controlled by its own config flag, i.e. not the generic `allow_hyperscan` one.

All in all, I think vectorscan has its merits for checking the existence of matches, but it isn't great for string processing scenarios like here. Imho, we can throw this PR away and just merge the first cleanup commit.