@nickitat 2022-11-02
in light of differences in behaviour you've mentioned the idea of this replacement is indeed questionable. but the case when we only need to know if a match exists (`where like '<str>'` et al) would be very cool to optimise if it is possible.