@rschu1ze 2022-11-02
@nickitat Yes, that should be possible and also worthy to optimize as `like` is quite popular. I created #42906 for that.

Perhaps you would like to take a look at #42907? It's the first commit of this PR and should be simple to review. Thanks.